extendable tooltip renderer
